优化PromptManager的封装,封装能够返回预热Prompt的接口
